pe/client/requirements.txt
websocket-client
smbus
RPi.GPIO
requests
picamera
